requirements.txt
aerich==0.6.3; python_version >= "3.7" and python_version < "4.0"
aiohttp==3.8.1; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0"
aioredis==1.3.1
aiosqlite==0.17.0; python_version >= "3.7" and python_version < "4.0"
apraw==0.6.9a0; python_version >= "3.6"
asyncdagpi==4.1.5; python_version >= "3.8" and python_version < "4.0"
asyncpg==0.26.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.0"
asyncpraw==7.5.0; python_version >= "3.6"
beautifulsoup4==4.11.1; python_full_version > "3.0.0"
black==22.8.0; python_full_version >= "3.6.2"
colorama==0.4.5; python_full_version >= "3.6.2" and python_version >= "3.4" and python_full_version < "4.0.0" and (python_version >= "2.7" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.5.0" and platform_system == "Windows") and (python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0") and (python_version >= "3.4" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.4")
download==0.3.5
fuzzywuzzy==0.18.0
genshinstats==1.4.11.3; python_version >= "3.7"
halo==0.0.31; python_version >= "3.4"
humanize==4.3.0; python_version >= "3.6"
inquirerpy==0.3.4; python_version >= "3.7" and python_version < "4.0"
jeyyapi @ git+https://github.com/JeyyGit/jeyyapi.git@main
jishaku @ git+https://github.com/Gorialis/jishaku@master ; python_full_version >= "3.8.0"
nekos.py==1.1.0
nudepy==0.5.1
numpy==1.23.2; python_version >= "3.7"
ossapi==2.3.0
osu-tools==0.2.13; python_version >= "3"
pillow==9.2.0; python_version >= "3.6"
pre-commit==2.20.0; python_full_version >= "3.6.1"
psutil==5.9.1; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
py-cord @ git+https://github.com/Pycord-Development/pycord.git@master ; python_full_version >= "3.8.0"
pycord18n @ git+https://github.com/YoungTrep/pycord18n.git@main
pycordutils==0.0.2; python_version >= "3.6"
pydantic==1.10.1; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
pynacl==1.5.0; python_full_version >= "3.8.0"
pypresence @ git+https://github.com/qwertyquerty/pypresence@master ; python_version >= "3.5"
pyshorteners==1.0.1; python_version >= "3.6"
python-anilist==1.0.9; python_version >= "3.6"
pytz==2022.2.1; python_version >= "3.7" and python_version < "4.0"
pyyaml==6.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
rich==12.5.1; python_full_version >= "3.6.2" and python_full_version < "4.0.0"
roblox==2.0.0; python_version >= "3.7"
statcord.py==3.1.0; python_version >= "3.5"
sympy==1.11.1; python_version >= "3.6"
tiktokapi==5.2.2
topggpy @ git+https://github.com/top-gg/python-sdk/@master ; python_version >= "3.6"
tortoise-orm @ git+https://github.com/tortoise/tortoise-orm/@develop; python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
virtualenv==20.16.4; python_full_version >= "3.6.1"
watchgod==0.8.2; python_version >= "3.5"
websockets==10.3; python_version >= "3.7"
yarl==1.8.1; python_version >= "3.6"
youtube-dl==2021.12.17
zipp==3.8.1; python_version >= "3.7"
